decimal values and percent on sales being used

+++ apps/app/lib/app/contracts/contract.ex
@@ -7,7 +7,7 @@ defmodule App.Contracts.Contract do
     belongs_to(:influencer, App.Influencers.Influencer)
     field(:minimum_points, :integer, default: 0)
     field(:payment_period, :integer, default: 0)
-    field(:points, :integer, default: 0)
+    field(:points, :decimal, default: 0.0)
     has_many(:voucher, App.Vouchers.Voucher)
 
     timestamps()
