+++ apps/app/priv/repo/migrations/20180410092403_create_contracts.exs
@@ -7,7 +7,7 @@ defmodule App.Repo.Migrations.CreateContracts do
       add(:brand_id, references(:brands, on_delete: :nothing))
       add(:minimum_points, :integer)
       add(:payment_period, :integer)
-      add(:points, :integer)
+      add(:points, :decimal)
 
       timestamps()
     end
